requirements.txt
streamlit
google-generativeai
fpdf
PyMuPDF
matplotlib
pandas
numpy